+++ policyengine_core/errors/__init__.py
@@ -1,6 +1,7 @@
 from .cycle_error import CycleError
 from .empty_argument_error import EmptyArgumentError
 from .nan_creation_error import NaNCreationError
+from .non_deterministic_formula_error import NonDeterministicFormulaError
 from .parameter_not_found_error import ParameterNotFoundError
 from .parameter_parsing_error import ParameterParsingError
 from .period_mismatch_error import PeriodMismatchError